Backend/llm/requirements.txt
google-genai
uvicorn
fastapi
